editoast: move train schedule allowance to schemas
