+++ editoast/editoast_schemas/src/lib.rs
@@ -1,1 +1,7 @@
 pub mod rolling_stock;
+pub mod train_schedule;
+
+editoast_common::schemas! {
+    rolling_stock::schemas(),
+    train_schedule::schemas(),
+}
