+++ editoast/src/core/simulation.rs
@@ -1,4 +1,5 @@
 use editoast_schemas::rolling_stock::RollingStockComfortType;
+use editoast_schemas::train_schedule::Allowance;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
 use utoipa::ToSchema;
@@ -10,7 +11,6 @@ use crate::models::train_schedule::Mrsp;
 use crate::models::train_schedule::RjsPowerRestrictionRange;
 use crate::models::train_schedule::SimulationPowerRestrictionRange;
 use crate::models::train_schedule::TrainScheduleOptions;
-use crate::models::Allowance;
 use crate::models::Pathfinding;
 use crate::models::PathfindingPayload;
 use crate::models::ResultTrain;
