+++ editoast/src/models/train_schedule.rs
@@ -7,6 +7,7 @@ use diesel::ExpressionMethods;
 use diesel::QueryDsl;
 use diesel_async::RunQueryDsl;
 use editoast_derive::Model;
+use editoast_schemas::train_schedule::Allowance;
 use serde::Deserialize;
 use serde::Serialize;
 use utoipa::ToSchema;
@@ -43,11 +44,6 @@ editoast_common::schemas! {
     ElectrificationUsage,
     ElectrificationRange,
     SimulationPowerRestrictionRange,
-    AllowanceValue,
-    AllowanceDistribution,
-    RangeAllowance,
-    EngineeringAllowance,
-    StandardAllowance,
     Allowance,
     ScheduledPoint,
     ResultTrain,
@@ -486,62 +482,6 @@ pub struct SimulationPowerRestrictionRange {
     stop: f64,
     code: String,
     handled: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
-#[serde(tag = "value_type")]
-pub enum AllowanceValue {
-    #[serde(rename = "time_per_distance")]
-    TimePerDistance { minutes: f64 },
-    #[serde(rename = "time")]
-    Time { seconds: f64 },
-    #[serde(rename = "percentage")]
-    Percent { percentage: f64 },
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
-#[serde(rename_all = "UPPERCASE")]
-pub enum AllowanceDistribution {
-    Mareco,
-    Linear,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
-pub struct RangeAllowance {
-    begin_position: f64,
-    end_position: f64,
-    value: AllowanceValue,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
-
-pub struct EngineeringAllowance {
-    #[serde(flatten)]
-    range: RangeAllowance,
-    distribution: AllowanceDistribution,
-    #[serde(default = "default_capacity_speed_limit")]
-    #[schema(default = default_capacity_speed_limit)]
-    capacity_speed_limit: f64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
-pub struct StandardAllowance {
-    default_value: AllowanceValue,
-    ranges: Vec<RangeAllowance>,
-    distribution: AllowanceDistribution,
-    #[serde(default = "default_capacity_speed_limit")]
-    capacity_speed_limit: f64,
-}
-
-const fn default_capacity_speed_limit() -> f64 {
-    -1.0
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
-#[serde(tag = "allowance_type", rename_all = "lowercase")]
-pub enum Allowance {
-    Engineering(EngineeringAllowance),
-    Standard(StandardAllowance),
 }
 
 #[derive(Debug, Clone, Derivative, Serialize, Deserialize, PartialEq, ToSchema)]
