+++ editoast/src/views/single_simulation.rs
@@ -3,6 +3,7 @@ use actix_web::web::Data;
 use actix_web::web::Json;
 use editoast_derive::EditoastError;
 use editoast_schemas::rolling_stock::RollingStockComfortType;
+use editoast_schemas::train_schedule::Allowance;
 use serde_derive::Deserialize;
 use serde_derive::Serialize;
 use thiserror::Error;
@@ -16,7 +17,6 @@ use crate::core::AsCoreRequest;
 use crate::core::CoreClient;
 use crate::error::InternalError;
 use crate::error::Result;
-use crate::models::train_schedule::Allowance;
 use crate::models::train_schedule::ElectrificationRange;
 use crate::models::train_schedule::Mrsp;
 use crate::models::train_schedule::ResultTrain;
