+++ editoast/src/views/stdcm/mod.rs
@@ -6,6 +6,7 @@ use actix_web::Responder;
 use derivative::Derivative;
 use editoast_derive::EditoastError;
 use editoast_schemas::rolling_stock::RollingStockComfortType;
+use editoast_schemas::train_schedule::AllowanceValue;
 use serde::Deserialize;
 use serde::Serialize;
 use thiserror::Error;
@@ -26,7 +27,6 @@ use crate::core::stdcm::STDCMCoreStep;
 use crate::core::AsCoreRequest;
 use crate::core::CoreClient;
 use crate::error::Result;
-pub use crate::models::train_schedule::AllowanceValue;
 use crate::models::Pathfinding;
 use crate::models::PathfindingChangeset;
 use crate::models::PathfindingPayload;
