+++ editoast/src/views/train_schedule/mod.rs
@@ -20,6 +20,7 @@ use diesel_async::AsyncConnection;
 use diesel_async::RunQueryDsl;
 use editoast_derive::EditoastError;
 use editoast_schemas::rolling_stock::RollingStockComfortType;
+use editoast_schemas::train_schedule::Allowance;
 use itertools::izip;
 use serde::Serialize;
 use serde_derive::Deserialize;
@@ -39,7 +40,6 @@ use crate::core::CoreClient;
 use crate::error::InternalError;
 use crate::error::Result;
 use crate::models::train_schedule::filter_invalid_trains;
-use crate::models::train_schedule::Allowance;
 use crate::models::train_schedule::RjsPowerRestrictionRange;
 use crate::models::train_schedule::ScheduledPoint;
 use crate::models::train_schedule::TrainScheduleOptions;
